ResponseFileGeneratorService is created.
